requirements.txt
alembic==1.4.3
aniso8601==8.1.0
attrs==20.3.0
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.1
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
typing-extensions==3.7.4.3
Werkzeug==0.16.1
zipp==3.4.0
